add transcript to url
