+++ package.json
@@ -8,6 +8,7 @@
     "@types/jquery": "^3.3.29",
     "@types/lodash": "^4.14.123",
     "@types/node": "12.6.8",
+    "@types/qs": "^6.9.1",
     "@types/rc-tooltip": "^3.7.1",
     "@types/react": "16.8.23",
     "@types/react-dom": "^16.8.5",
@@ -29,6 +30,7 @@
     "node-sass": "^4.13.0",
     "oncokb-styles": "^0.1.1",
     "prettier": "^1.18.2",
+    "qs": "^6.9.3",
     "rc-tooltip": "^3.7.3",
     "react": "^16.8.6",
     "react-bootstrap": "^1.0.0-beta.6",
